refactor(clustering): move the Leiden search out of CallGraph into the engine
Third of the stack. Pure move — the algorithm is byte-for-byte what it was.

The search never needed the graph: it reads an exported nx.DiGraph and a
delimiter, nothing else. Eight private CallGraph methods become module functions
in clustering/engine.py behind cluster_graph(), and CallGraph.cluster() becomes a
four-line call that still owns the cache.

detect_communities goes with them. It was a pass-through to
leiden_utils.find_partition, and leaving a clustering entry point behind on a
module that no longer clusters would defeat the point; cluster_helpers and the
determinism test now call find_partition directly. That test was named for the
wrapper, so it is renamed for what it actually exercises.

The partition is still cached on CallGraph — that moves next.

graph.py drops from 686 to 447 lines.
